Extract EmptyStateNoAccess to a separate cmp
